Safety snapshot of partial xs file split
By using #ifdef's and #defines such as IN_POGL_GL_XS, I am
segmenting the original OpenGL.xs file by the planned final
XS file location.  Once all the partitioning #ifdef--#endif
pairs are in, we'll copy the file to the new .xs names and
set only the specific #define for that file.  When the split
version is debugged and working we will remove the redundant
code sections.