OpenGL.xs has partitions...next to split to files.
The #ifdef IN_POGL_XXX_XS dividers are in place in
OpenGL.xs and it should be ready to create the actual
split out XS files.  Note: I discovered that you cannot
span a MODULE directive line in an XS file with #ifdef
to #endif (or #else).  Either side works ok.