Renamed pogl_main.xs to pogl_gl.xs
This is the planned final name for the file
making the GL bindings.  It currently has
everything that has not been split out.
Similarly, pogl_rpn.xs contains the OGA
stuff due to the dependencies in the current
code.  Once the RPN declarations have been
refactored into an include file, the OGA
stuff will move to pogl_array.xs.