Revert to glutInitDisplayMode() in test.pl
glutInitDisplayString() is not fully implemented
in FreeGLUT so it is not possible to request
either an alpha channel buffer or one without
an alpha channel.  It may be possible to work
around with a try-and-die hack which would
try a couple of test applications to see which
end up running.