Update readme files and add GIT_CHANGES
This is an update of several of the "readme" files and
the introduction of the GIT_CHANGES file which is generated
for each release from the output of git log --stat command.

Once 0.58 is released to CPAN and the PDL POGL refactoring
is implemented we need a complete rewrite pass to correct
and clean up the documentation for this module to match the
reality of the current approach and code.