Don't remove HAVE_GLX define with *GLUT on linux
As part of an attempted cleanup of configuration output,
HAVE_GLX was being deleted on non-cygwin systems using
GLUT or FreeGLUT.  This issue will be addressed in a planned
refactoring of the POGL configuration info to make explicit
things the build system *has* versus the things the build
system *used* to make POGL, e.g. use FreeGLUT but has GLX also.