Add HAVE_FREEGLUT to skip if GLUT only
Problems building POGL have been reported for systems
when GLUT is being used instead of FreeGLUT for two reasons:
(1) FreeGLUT is not installed in which case we would like the
build to work but perhaps with less capability, or (2) both
GLUT and FreeGLUT are available but the wrong include file is
being used (i.e. glut.h instead of freeglut.h).