Placeholder for freeglut.dll.bat problem
This is a proposed fix for the current use of EXE_FILES in Makefile.PL
to install freeglut.dll on win32 machines.  EXE_FILES is for perl
scripts and not binaries.  The fix does work but it has the problem
of performing the install at configure time.  The whole process needs
to be reworked to happen at install time.