Apply fix to prevent freeglut.dll.bat install
The new code is probably better than doing an install of
freeglut.dll to some systemwide location.  If we can verify
that the paths are set up correctly and the code works, we
should remove the alternate install since it is less
problematic to keep libraries installed for perl in the
perl locations rather than global install locations.