Update TODO with PM on OS/2 info
We'll be discontinuing support of PM on OS/2 for GLX since
that platform does not appear to be in use.  This may be
moot since we have already moved from GLX to FreeGLUT for
windows.  Now only need to add the ability to use OpenGL
routines when passed appropriate platform contexts (GLX,WGL,AGL).