Add t/10_opengl_array.t fails to KNOWN_PROBLEMS
The TODO tests fails to ignore test failures when the
failure results in a crash of the interpreter itself.
Adding this fact to KNWON_PROBLEMS.  While the OpenGL::Array
problems are being debugged, maybe we could come up with a
way to make TODO skips work through crashes---perhaps by
spawning a separate interpreter for the tests?