new(): fix item_count calculation
new() "[c]reates an empty array object of $count rows made up [of] data
types @types" (see Array.pod).

Memory allocation for the array work as described (and calculation of
total_types_width work as expected), but item_count reflects the number
of rows, while other functions such as retrieve() and affine() interpret
it as the number of items (= values in the array, rather than rows in
the array).

This patch makes item_count reflect the number of items in the array, so
that functions like retrieve() and affine() work on the whole array.