synchronize freeglut bin/include with 2.8.1 from transmissionzero
The freeglut files were in a bit of a weird state, with version info
missing and the binaries having been overwritten at one point, without the
includes having been updated, so this commit synchs that to a known and
publicly accessible release.

As for the glut_util change, i'm not entirely sure why i did that, but i
found it in another commit i had made where i had experimentally updated
freeglut. Let's hope it helps.